Make runtime asserts into static asserts.
This change looks like the result of a compiler issue. Accessing
a static constexpr member using the dot operator should be an integral
constant expression (the base object should be ignored in this context).
For some reason, Clang believes otherwise. Perhaps this is a language
defect.

Referring to the order through the type works.